requirements.txt
GitPython
google-generativeai
python-dotenv
tomli